refactor(admin): extract Users tab into pages/admin/users.vue
